:recycle: 거래 완료 횟수 조회 메서드 래퍼 타입으로 변경

+++ server/wonprice/src/main/java/main/wonprice/domain/member/dto/MemberResponseDto.java
@@ -17,9 +17,9 @@ public class MemberResponseDto {
 
     private String phone;
 
-    private int postCount;
+    private Long postCount;
 
-    private int tradeCount;
+    private Long tradeCount;
 
     private MemberPicture picture;
 }
